Split up api_app tests. Added tests for field serialization on Change model.
